Update decorators tests
